+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
 - Enh #362: Template - Rendering & Caching
 - Fix #365: Fix allowed properties for Twig v3.14.1+
 - Enh #367: Migrate all element contents to single table
+- Fix #376: Use template name for allowed templates in the container element dynamic attributes
 
 1.10.11 (Unreleased)
 --------------------------